Remove broken buttons in Shipments page
These buttons are probably legacy code. I run the test suite
with older Solidus versions (v2.5, v2.4, v2.3) and still they
did not work. A spec existed for this feature, but it did not
verify that it actually works.

This functionality, however, exists (in working conditions and
is implemented in Solidus backend codebase) in the admin `Cart`
page,so I think it makes sense removing it altogether from the
`Shipments` page.