Update tests to avoid using finalize!
We should use complete! now.

Also, the transition to cart to complete is no more needed.
In the majority of the specs it's not needed to have the order
completed, so we can remove the finalize! call. In other cases
we can just mock the value of the completed? method.